更新至XCode6适配 修正了Order到History，History到Order需要Login中转，现在为直接展示 TODO：多次选择某商品时，价格计算会出现错误

+++ sterminal/BFActiveViewController.m
@@ -108,7 +108,7 @@ static NSString *GarbageString = @"Thread was being aborted.";
 {
     self.rString = [[NSMutableString alloc] init];
     //[textField resignFirstResponder];
-    self.activeLabelIndex = textField.tag - 2000;
+    self.activeLabelIndex = (int)textField.tag - 2000;
     self.currentTextField = (UITextField *)[self.textFields objectAtIndex:self.activeLabelIndex];
     
     NSLog(@"Active Index: %d", self.activeLabelIndex);
